분산락을 이용한 좋아요 수 갱신 유실 문제 해결 / Redis 를 이용한 중복 요청 방지

+++ src/main/java/com/example/codebase/domain/post/service/PostService.java
@@ -1,5 +1,6 @@
 package com.example.codebase.domain.post.service;
 
+import com.example.codebase.annotation.DistributedLock;
 import com.example.codebase.controller.dto.PageInfo;
 import com.example.codebase.domain.member.entity.Member;
 import com.example.codebase.domain.member.exception.NotFoundMemberException;
@@ -7,10 +8,11 @@ import com.example.codebase.domain.member.repository.MemberRepository;
 import com.example.codebase.domain.post.dto.*;
 import com.example.codebase.domain.post.entity.*;
 import com.example.codebase.domain.post.repository.PostCommentRepository;
-import com.example.codebase.domain.post.repository.PostDocumentRepository;
 import com.example.codebase.domain.post.repository.PostLikeMemberRepository;
 import com.example.codebase.domain.post.repository.PostRepository;
+import com.example.codebase.exception.LikePostDuplicatedRequestException;
 import com.example.codebase.s3.S3Service;
+import com.example.codebase.util.RedisUtil;
 import com.example.codebase.util.SecurityUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -19,6 +21,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -31,27 +34,31 @@ public class PostService {
     private final PostLikeMemberRepository postLikeMemberRepository;
     private final PostCommentRepository postCommentRepository;
 
-//    private final PostDocumentRepository postDocumentRepository;
+    //    private final PostDocumentRepository postDocumentRepository;
     private final S3Service s3Service;
 
+    private final RedisUtil redisUtil;
+
     @Autowired
+
     public PostService(
             PostRepository postRepository,
             MemberRepository memberRepository,
             PostLikeMemberRepository postLikeMemberRepository,
             PostCommentRepository postCommentRepository,
-            S3Service s3Service) {
+            S3Service s3Service, RedisUtil redisUtil) {
         this.postRepository = postRepository;
         this.memberRepository = memberRepository;
         this.postLikeMemberRepository = postLikeMemberRepository;
         this.postCommentRepository = postCommentRepository;
         this.s3Service = s3Service;
+        this.redisUtil = redisUtil;
     }
 
     @Transactional
     public PostResponseDTO createPost(PostCreateDTO postCreateDTO, String loginUsername) {
         Member author =
-            memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
+                memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
 
         Post newPost = Post.of(postCreateDTO, author);
         author.addPost(newPost);
@@ -82,7 +89,7 @@ public class PostService {
         PageInfo pageInfo = PageInfo.of(page, size, posts.getTotalPages(), posts.getTotalElements());
 
         List<PostResponseDTO> dtos =
-            posts.stream().map(PostResponseDTO::from).collect(Collectors.toList());
+                posts.stream().map(PostResponseDTO::from).collect(Collectors.toList());
         return PostsResponseDTO.of(dtos, pageInfo);
     }
 
@@ -92,21 +99,21 @@ public class PostService {
         PageRequest pageRequest = PageRequest.of(page, size, sort);
 
         Member member =
-            memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
+                memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
 
         Page<PostWithIsLiked> postPages = postRepository.findAllWithIsLiked(member, pageRequest);
         PageInfo pageInfo =
-            PageInfo.of(page, size, postPages.getTotalPages(), postPages.getTotalElements());
+                PageInfo.of(page, size, postPages.getTotalPages(), postPages.getTotalElements());
 
         List<PostResponseDTO> dtos =
-            postPages.stream().map(PostResponseDTO::from).collect(Collectors.toList());
+                postPages.stream().map(PostResponseDTO::from).collect(Collectors.toList());
         return PostsResponseDTO.of(dtos, pageInfo);
     }
 
     @Transactional
     public PostResponseDTO updatePost(Long postId, PostUpdateDTO postUpdateDTO, String loginUsername) {
         Post post =
-            postRepository.findById(postId).orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
+                postRepository.findById(postId).orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
 
         if (!SecurityUtil.isAdmin() && !post.getAuthor().getUsername().equals(loginUsername)) {
             throw new RuntimeException("게시글 작성자만 수정할 수 있습니다.");
@@ -119,7 +126,7 @@ public class PostService {
     @Transactional
     public void deletePost(Long postId, String loginUsername) {
         Post post =
-            postRepository.findById(postId).orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
+                postRepository.findById(postId).orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
 
         if (!SecurityUtil.isAdmin() && !post.getAuthor().getUsername().equals(loginUsername)) {
             throw new RuntimeException("게시글 작성자만 삭제할 수 있습니다.");
@@ -136,16 +143,16 @@ public class PostService {
 
     public PostWithLikesResponseDTO getPost(Long postId) {
         Post post =
-            postRepository.findById(postId).orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
+                postRepository.findById(postId).orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
 
         post.incressView();
 
         List<PostLikeMemberDTO> postLikeMemberDtos =
-            post.getPostLikeMembers().stream()
-                .map(
-                    (PostLikeMember likeMember) ->
-                        PostLikeMemberDTO.of(likeMember.getMember(), likeMember.getLikedTime()))
-                .collect(Collectors.toList());
+                post.getPostLikeMembers().stream()
+                        .map(
+                                (PostLikeMember likeMember) ->
+                                        PostLikeMemberDTO.of(likeMember.getMember(), likeMember.getLikedTime()))
+                        .collect(Collectors.toList());
 
         return PostWithLikesResponseDTO.create(post, postLikeMemberDtos);
     }
@@ -155,48 +162,62 @@ public class PostService {
         PostWithLikesResponseDTO post = getPost(postId);
 
         Member member =
-            memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
+                memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
 
         postLikeMemberRepository
-            .findById(PostLikeMemberIds.of(member, post.getId()))
-            .ifPresent((PostLikeMember likeMember) -> post.setIsLiked(true));
+                .findById(PostLikeMemberIds.of(member, post.getId()))
+                .ifPresent((PostLikeMember likeMember) -> post.setIsLiked(true));
 
         return post;
     }
 
+//    @Transactional(isolation = Isolation.SERIALIZABLE) //
+    @DistributedLock(key = "#postId")
     @Transactional
     public PostResponseDTO likePost(Long postId, String loginUsername) {
+        if (isDuplicatedRequest(postId, loginUsername)) {
+            throw new LikePostDuplicatedRequestException();
+        }
+        duplicatedRequestSave(postId, loginUsername);
+
         Post post =
-            postRepository.findById(postId).orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
+                postRepository.findById(postId).orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
 
         Member member =
-            memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
+                memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
 
         Optional<PostLikeMember> likeMember =
-            postLikeMemberRepository.findById(PostLikeMemberIds.of(member, post));
+                postLikeMemberRepository.findById(PostLikeMemberIds.of(member, post));
 
         boolean isLiked = likeMember.isPresent();
         if (isLiked) {
             postLikeMemberRepository.delete(likeMember.get());
+            post.decreaseLike();
         } else {
             PostLikeMember newLike = PostLikeMember.of(post, member);
-            post.addLikeMember(newLike);
+            post.incressLike();
+            postLikeMemberRepository.save(newLike);
         }
-
-        Integer likeCount = postLikeMemberRepository.countByPostId(post.getId());
-        post.setLikes(likeCount);
 
         return PostResponseDTO.of(post, !isLiked);
     }
 
+    private boolean isDuplicatedRequest(Long postId, String username) {
+        return redisUtil.getData(postId + "_" + username).isPresent();
+    }
+
+    private void duplicatedRequestSave(Long postId, String username) {
+        redisUtil.setDataAndExpire(postId + "_" + username, String.valueOf(LocalDateTime.now()), 3000);
+    }
+
     @Transactional
     public PostResponseDTO createComment(
-        Long postId, PostCommentCreateDTO commentCreateDTO, String loginUsername) {
+            Long postId, PostCommentCreateDTO commentCreateDTO, String loginUsername) {
         Post post =
-            postRepository.findById(postId).orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
+                postRepository.findById(postId).orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
 
         Member author =
-            memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
+                memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
 
         PostComment newComment = PostComment.of(commentCreateDTO, author);
         setParentComment(commentCreateDTO, post, newComment);
@@ -212,9 +233,9 @@ public class PostService {
         if (parentCommentId.isPresent()) { // 대댓글 일 시
             Long perentId = parentCommentId.get();
             PostComment parentComment =
-                postCommentRepository
-                    .findByIdAndPost(perentId, post)
-                    .orElseThrow(() -> new RuntimeException("존재하지 않는 댓글이거나 해당 게시글에 속해있지 않습니다."));
+                    postCommentRepository
+                            .findByIdAndPost(perentId, post)
+                            .orElseThrow(() -> new RuntimeException("존재하지 않는 댓글이거나 해당 게시글에 속해있지 않습니다."));
 
             if (parentComment.getParent() != null) { // mention
                 newComment.setParent(parentComment.getParent());
@@ -227,14 +248,14 @@ public class PostService {
 
     @Transactional
     public PostResponseDTO updateComment(
-        Long commentId, PostCommentUpdateDTO commentUpdateDTO, String loginUsername) {
+            Long commentId, PostCommentUpdateDTO commentUpdateDTO, String loginUsername) {
         PostComment comment =
-            postCommentRepository
-                .findById(commentId)
-                .orElseThrow(() -> new RuntimeException("존재하지 않는 댓글입니다."));
+                postCommentRepository
+                        .findById(commentId)
+                        .orElseThrow(() -> new RuntimeException("존재하지 않는 댓글입니다."));
 
         Member author =
-            memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
+                memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
 
         if (!SecurityUtil.isAdmin() && !comment.getAuthor().equals(author)) {
             throw new RuntimeException("댓글 작성자만 수정할 수 있습니다.");
@@ -248,12 +269,12 @@ public class PostService {
     @Transactional
     public PostResponseDTO deleteComment(Long commentId, String loginUsername) {
         PostComment comment =
-            postCommentRepository
-                .findById(commentId)
-                .orElseThrow(() -> new RuntimeException("존재하지 않는 댓글입니다."));
+                postCommentRepository
+                        .findById(commentId)
+                        .orElseThrow(() -> new RuntimeException("존재하지 않는 댓글입니다."));
 
         Member author =
-            memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
+                memberRepository.findByUsername(loginUsername).orElseThrow(NotFoundMemberException::new);
 
         if (!SecurityUtil.isAdmin() && !comment.getAuthor().equals(author)) {
             throw new RuntimeException("댓글 작성자만 삭제할 수 있습니다.");
